refactor(speakers): :recycle: share diarization session init + centroid persistence across pipelines
